Replace CirisSpec with CirisSuite tests
